refactor: EmotionRecordDetailView를 별도 파일로 분리하여 구조 개선
- 파일 구조 개선으로 유지보수성 향상
